requirements.txt
streamlit
pandas
pdfplumber
python-docx
scikit-learn
numpy
supabase
